+++ src/command_modules/azure-cli-storage/azure/cli/command_modules/storage/tests/latest/test_storage_account_scenarios.py
@@ -3,7 +3,7 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 from azure.cli.testsdk import (ScenarioTest, JMESPathCheck, ResourceGroupPreparer,
-                               StorageAccountPreparer, api_version_constraint)
+                               StorageAccountPreparer, api_version_constraint, live_only)
 from azure.cli.core.profiles import ResourceType
 from .storage_test_util import StorageScenarioMixin
 
@@ -117,8 +117,8 @@ class StorageAccountTests(StorageScenarioMixin, ScenarioTest):
 
         self.cmd('storage account show-connection-string -g {} -n {} --protocol http'.format(
             resource_group, name), checks=[
-            JMESPathCheck("contains(connectionString, 'https')", False),
-            JMESPathCheck("contains(connectionString, '{}')".format(name), True)])
+                JMESPathCheck("contains(connectionString, 'https')", False),
+                JMESPathCheck("contains(connectionString, '{}')".format(name), True)])
 
         self.cmd('storage account update -g {} -n {} --tags foo=bar cat'
                  .format(resource_group, name),