+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -2,14 +2,12 @@
 
 Release History
 ===============
-0.2.4
-+++++
-* `webapp/functionapp identity show`: exception handling to exit with code 3 upon a missing resource for consistency
 
 0.2.3
 +++++
 * support CORS on functionapp & webapp
 * arm tag support on create commands
+* `webapp/functionapp identity show`: exception handling to exit with code 3 upon a missing resource for consistency
 
 0.2.2
 +++++